// SpringBoot - Invoices_Tracker/InvoiceTrackerFE/src/shared/models/user.interface.ts
export interface Users{
    firstName: string;
    lastName: string;
    email: string;
    address: string;
    role: number;
    age: number;
    password: string;
    id?: number;
}
export interface UsersInterface{
    firstName: string;
    lastName: string;
    email: string;
    address: string;
    role: number;
    age: number;
    password: string;
    id?: number;
}